requirements.txt
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
asyncio
pandas==2.1.4
numpy==1.24.3
Pillow==10.1.0
python-dateutil==2.8.2
